# my-nest-crud/swagger-spec.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"operationId":"AuthController_login","summary":"Realizar login com nome de usuário e senha","parameters":[],"requestBody":{"required":true,"description":"Dados para realizar o login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginDto"}}}},"responses":{"200":{"description":"Login bem-sucedido","content":{"application/json":{"schema":{"example":{"access_token":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}}}}},"401":{"description":"Credenciais inválidas"}},"tags":["auth"]}},"/auth/register":{"post":{"operationId":"AuthController_register","summary":"Registrar um novo usuário","parameters":[],"requestBody":{"required":true,"description":"Dados para registrar um novo usuário","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRegisterDto"}}}},"responses":{"201":{"description":"Registro bem-sucedido","content":{"application/json":{"schema":{"example":{"username":"john_doe","password":"******"}}}}},"400":{"description":"Dados inválidos para registro"}},"tags":["auth"]}},"/auth/profile":{"get":{"operationId":"AuthController_getProfile","summary":"Obter o perfil do usuário autenticado","parameters":[],"responses":{"200":{"description":"Perfil do usuário autenticado","content":{"application/json":{"schema":{"example":{"userId":1,"username":"john_doe"}}}}}},"tags":["auth"],"security":[{"bearer":[]}]}},"/books":{"post":{"operationId":"BooksController_create","summary":"Criar um novo livro","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookDto"}}}},"responses":{"201":{"description":"Livro criado com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookDto"}}}},"400":{"description":"Dados inválidos."}},"tags":["books"],"security":[{"bearer":[]}]},"get":{"operationId":"BooksController_findAll","summary":"Listar todos os livros","parameters":[],"responses":{"200":{"description":"Lista de livros retornada com sucesso.","example":[{"title":"Clean Code","author":"Robert C. Martin","description":"Um livro sobre boas práticas de desenvolvimento de software.","publishedDate":"2021-06-15T00:00:00.000Z"},{"title":"The Pragmatic Programmer","author":"Andrew Hunt","description":"A journey to mastery in programming.","publishedDate":"2019-10-20T00:00:00.000Z"}],"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CreateBookDto"}},"example":[{"title":"Clean Code","author":"Robert C. Martin","description":"Um livro sobre boas práticas de desenvolvimento de software.","publishedDate":"2021-06-15T00:00:00.000Z"},{"title":"The Pragmatic Programmer","author":"Andrew Hunt","description":"A journey to mastery in programming.","publishedDate":"2019-10-20T00:00:00.000Z"}]}}}},"tags":["books"],"security":[{"bearer":[]}]}},"/books/{id}":{"get":{"operationId":"BooksController_findOne","summary":"Obter detalhes de um livro específico","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Livro encontrado com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookDto"}}}}},"tags":["books"],"security":[{"bearer":[]}]},"put":{"operationId":"BooksController_update","summary":"Atualizar informações de um livro","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookDto"}}}},"responses":{"200":{"description":"Livro atualizado com sucesso.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookDto"}}}}},"tags":["books"],"security":[{"bearer":[]}]},"delete":{"operationId":"BooksController_remove","summary":"Deletar um livro","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Livro deletado com sucesso."}},"tags":["books"],"security":[{"bearer":[]}]}}},"info":{"title":"Book API","description":"CRUD de Livros","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"AuthLoginDto":{"type":"object","properties":{"username":{"type":"string","description":"Nome de usuário para login","example":"john_doe"},"password":{"type":"string","description":"Senha do usuário","example":"password123"}},"required":["username","password"]},"AuthRegisterDto":{"type":"object","properties":{"username":{"type":"string","description":"Nome de usuário para registro","example":"john_doe"},"password":{"type":"string","description":"Senha do usuário para registro","example":"password123"}},"required":["username","password"]},"CreateBookDto":{"type":"object","properties":{"title":{"type":"string","description":"Título do livro","example":"Clean Code"},"author":{"type":"string","description":"Autor do livro","example":"Robert C. Martin"},"description":{"type":"string","description":"Descrição do livro","example":"Um livro sobre boas práticas de desenvolvimento de software."},"publishedDate":{"format":"date-time","type":"string","description":"Data de publicação","example":"2021-06-15T00:00:00.000Z"}},"required":["title","author"]}}}}